requirements.txt
msoffcrypto-tool
pandas
